organize some configs because god this is messy
