+++ src/configs/networking/axios.config.ts
@@ -7,8 +7,7 @@ import axios, {
 	InternalAxiosRequestConfig,
 } from 'axios'
 import { nitroFetchOnWorklet } from 'react-native-nitro-fetch'
-import { captureError } from '../utils/logging'
-import LoggingContext from '../utils/logging/enums'
+import { NETWORK_TIMEOUT } from './thresholds.config'
 
 type NitroMappedResponse = {
 	status: number
@@ -178,7 +177,7 @@ function buildFullURL(config: InternalAxiosRequestConfig): string {
 }
 
 const AXIOS_INSTANCE = axios.create({
-	timeout: 60000,
+	timeout: NETWORK_TIMEOUT,
 	adapter: nitroAxiosAdapter,
 })
