+++ src/configs/networking/thresholds.config.ts
@@ -1,0 +1,4 @@
+/**
+ * Default network timeout in ms
+ */
+export const NETWORK_TIMEOUT = 30000
